backend/requirements.txt
fastapi
uvicorn
python-multipart
langchain-core
langchain-openai
langchain-community
langgraph
pydantic
pydantic[email]
httpx
pymongo
python-dotenv
chromadb
requests
beautifulsoup4
pypdf
python-docx
trafilatura
tiktoken
anthropic
boto3
aiolimiter
prometheus-client
passlib
argon2-cffi
PyJWT
opentelemetry-api
opentelemetry-sdk
opentelemetry-exporter-otlp
rq
redis
feedparser
apscheduler
